Back to int keys to simplify sorting; errno field for marker is 32-bit in any case

+++ clients/drcachesim/tools/syscall_mix.cpp
@@ -127,8 +127,7 @@ syscall_mix_t::parallel_shard_memref(void *shard_data, const memref_t &memref)
         ++shard->stats.syscall_trace_counts[syscall_num];
     } else if (memref.marker.type == TRACE_TYPE_MARKER &&
                memref.marker.marker_type == TRACE_MARKER_TYPE_SYSCALL_FAILED) {
-        ++shard->stats
-              .syscall_errno_counts[static_cast<int64_t>(memref.marker.marker_value)];
+        ++shard->stats.syscall_errno_counts[static_cast<int>(memref.marker.marker_value)];
     }
     return true;
 }
@@ -190,7 +189,7 @@ syscall_mix_t::print_results()
     if (!total.syscall_errno_counts.empty()) {
         std::cerr << std::setw(15) << "failure count"
                   << " : " << std::setw(9) << "failure code\n";
-        std::vector<std::pair<int64_t, int64_t>> sort_errno(
+        std::vector<std::pair<int, int64_t>> sort_errno(
             total.syscall_errno_counts.begin(), total.syscall_errno_counts.end());
         std::sort(sort_errno.begin(), sort_errno.end(), cmp_second_val);
         for (const auto &keyvals : sort_errno) {
